separate style overrides out of stepper component, update button to Link
