+++ components/TestLibraryStepper/TestLibraryStepper.tsx
@@ -1,11 +1,9 @@
-﻿import React, {useState} from "react";
-import {Step, StepConnector, StepIconProps, StepLabel, Stepper, Typography} from '@material-ui/core';
-import {createMuiTheme, makeStyles, MuiThemeProvider} from '@material-ui/core/styles';
-import Router from "next/router";
-import {withStyles} from "@material-ui/styles";
-import clsx from "clsx";
-import {Check, MoreHoriz} from "@material-ui/icons";
+﻿import React from "react";
+import {Step, StepLabel, Stepper, Typography} from '@material-ui/core';
+import {MuiThemeProvider} from '@material-ui/core/styles';
 import {CandidateTestStatus} from "../../pages/api/candidateApiClient.module";
+import {TestSwitchConnector, TestSwitchStepIcon, testSwitchTheme} from "../TestLibraryOverrides/TestLibraryOverrides"
+import Link from "next/link";
 
 interface TestLibraryStepperProps {
     candidateTestStatus: CandidateTestStatus[];
@@ -28,86 +26,6 @@ export default function TestLibraryStepper(props: TestLibraryStepperProps): JSX.
     //TODO this is example data, set active step with candidate number of results, maybe props.candidateTestStatus.results.length?
     const activeStep = 2;
 
-    //overrides for material ui stepper styles
-    const testSwitchTheme = createMuiTheme({
-        palette: {
-            primary: {
-                main: 'rgb(255, 193, 0)'
-            }
-        },
-    });
-
-    //overrides for material ui icon styles
-    const TestSwitchIconStyles = makeStyles({
-        root: {
-            backgroundColor: '#eaeaf0',
-            zIndex: 1,
-            color: '#fff',
-            fontSize: 18,
-            width: 30,
-            height: 30,
-            display: 'flex',
-            borderRadius: '50%',
-            justifyContent: 'center',
-        },
-        active: {
-            color: 'rgb(255, 193, 0)',
-            fontSize: 18,
-        },
-        current: {
-            color: 'rgb(255, 255, 255)',
-            zIndex: 0,
-            fontSize: 18,
-            width: 30,
-            height: 30,
-            borderRadius: '50%',
-            backgroundColor: 'rgb(255, 193, 0)',
-        },
-        completed: {
-            color: 'rgb(255, 255, 255)',
-            zIndex: 0,
-            fontSize: 18,
-            width: 30,
-            height: 30,
-            borderRadius: '50%',
-            backgroundColor: 'rgb(255, 193, 0)',
-        },
-    });
-
-    function TestSwitchStepIcon(props: StepIconProps) {
-        const classes = TestSwitchIconStyles();
-        const {active, completed} = props;
-        return (
-            <div
-                className={clsx(classes.root, {
-                    [classes.active]: active,
-                })}
-            >
-                {completed ? <Check className={classes.completed}/> : <MoreHoriz className={classes.current}/>}
-            </div>
-        );
-    }
-
-    //overrides for material ui stepper connector styles
-    const TestSwitchConnector = withStyles({
-        active: {
-            '& $line': {
-                borderColor: 'rgb(255, 193, 0)',
-            },
-        },
-        completed: {
-            '& $line': {
-                borderColor: 'rgb(255, 193, 0)',
-            },
-        },
-        line: {
-            borderColor: '#eaeaf0',
-            borderTopWidth: 4,
-            borderRadius: 1,
-            zIndex: -2,
-            lineHeight: 50,
-        },
-    })(StepConnector);
 
     return (
         <div className="stepperContainer">
@@ -131,11 +49,9 @@ export default function TestLibraryStepper(props: TestLibraryStepperProps): JSX.
                     <div className="stepperBtnContainer">
                         <div>
                             <Typography align={"center"}>
-                                <a className="stepperBtn" href="#"
-                                   onClick={() => Router.push(`/test${activeStep + 1}`)}
-                                >
+                                <Link href={`/test${activeStep + 1}`}><a className="stepperBtn">
                                     {activeStep === steps.length - 1 ? 'Start Final Test' : `Start Test ${activeStep + 1}`}
-                                </a>
+                                </a></Link>
                             </Typography>
                         </div>
                     </div>
